requirements.txt
opencv-python-headless>=4.8.0
numpy>=1.20.0
streamlit>=1.28.0
Pillow>=8.0.0
streamlit-webrtc>=0.47.6
av>=12.3.0